app/requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
astunparse==1.6.3
audioread==3.0.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
decorator==5.1.1
distro==1.9.0
Django==4.2.3
django-cors-headers==4.3.1
djangorestframework==3.15.1
filelock==3.14.0
flatbuffers==24.3.25
fsspec==2024.5.0
gast==0.5.4
google-pasta==0.2.0
grpcio==1.64.0
gunicorn==22.0.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.2
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.4
joblib==1.4.2
keras==3.3.3
keybert==0.8.4
kiwipiepy==0.17.1
kiwipiepy-model==0.17.0
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.42.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mkl==2021.4.0
ml-dtypes==0.3.2
mpmath==1.3.0
msgpack==1.0.8
mysqlclient==2.2.4
namex==0.0.8
networkx==3.3
numba==0.59.1
numpy==1.26.4
openai==1.30.4
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pillow==10.3.0
platformdirs==4.2.2
pooch==1.8.1
protobuf==4.25.3
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.7.2
pydantic_core==2.18.3
pydub==0.25.1
Pygments==2.18.0
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.2
rich==13.7.1
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
sentence-transformers==3.0.0
setuptools==70.0.0
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
soxr==0.3.7
sqlparse==0.5.0
sympy==1.12
tbb==2021.12.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
termcolor==2.4.0
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.0
tqdm==4.66.4
transformers==4.41.1
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
